Assorted patches:
 ------  BUILD PROCESS  ------

 Title:  "add utilities to make test dependencies"
  From:  Robin Barker <rmb1@cise.npl.co.uk>
Msg-ID:  <2607.9805211303@tempest.cise.npl.co.uk>
 Files:  Makefile.SH

 Title:  "Add 'make nok' complement to 'make ok'"
  From:  "M.J.T. Guy" <mjtg@cus.cam.ac.uk>
Msg-ID:  <E0ycRDf-0005Wh-00@taurus.cus.cam.ac.uk>
 Files:  Makefile.SH

 Title:  "further h2ph patches (add enum support)"
  From:  Billy <wdconsta@cs.adelaide.edu.au>
Msg-ID:              <Pine.SV4.3.93.980521025541.14577A-100000@xenon.teaching.cs.adelaide.edu.au>
 Files:  MANIFEST t/lib/h2ph.pht t/lib/h2ph.t utils/h2ph.PL

 ------  CORE LANGUAGE  ------

 Title:  "Fix %! error spelling and add perldiag.pod entry"
  From:  Graham Barr <gbarr@pobox.com>, Tim Bunce
Msg-ID:  <19980524193101.A573@pobox.com>
 Files:  pod/perldiag.pod gv.c

 Title:  "Remove obsolete Win32 uppercasing ENV code"
  From:  Gurusamy Sarathy <gsar@engin.umich.edu>
Msg-ID:  <199805201510.LAA28676@aatma.engin.umich.edu>
 Files:  perl.c

 Title:  "Don't mung $! on implicit close"
  From:  Chip Salzenberg <chip@perl.org>
Msg-ID:  <19980525113309.A15845@perlsupport.com>
 Files:  doio.c

 Title:  "Maint trial 3 fails on SunOS 4.1.3 with Sun cc"
  From:  Andy Dougherty <doughera@lafcol.lafayette.edu>
Msg-ID:  <Pine.SUN.3.96.980527113114.26608D-100000@newton.phys>
 Files:  doio.c

 ------  DOCUMENTATION  ------

 Title:  "doc patch: you canna return an array ( list context: || vs or)"
  From:  Jarkko Hietaniemi <jhi@cc.hut.fi>
Msg-ID:  <oeeemxguf5h.fsf_-_@alpha.hut.fi>
 Files:  pod/perldebug.pod pod/perlfunc.pod pod/perltie.pod pod/perltrap.pod

 Title:  "doc patch: @ needs escaping in m/\Q\E/ environment"
  From:  "M.J.T. Guy" <mjtg@cus.cam.ac.uk>
Msg-ID:  <E0yecim-0002qr-00@taurus.cus.cam.ac.uk>
 Files:  pod/perlop.pod pod/perlre.pod

 Title:  "Discrepancy between perlop.pod and m// operator", "Doc fix: Only
         with /g does list context get matches without parens"
  From:  Greg Chapman <glc@well.com>, Tom Christiansen
         <tchrist@jhereg.perl.com>, Tom Phoenix
         <rootbeer@teleport.com>
Msg-ID:  <000201bd865e$f3bf72e0$1f04400c@assigned.well.com>,
         <199805231559.JAA21316@jhereg.perl.com>,
         <Pine.GSO.3.96.980523084947.22348I-100000@user2.teleport.com>
 Files:  pod/perlop.pod

 Title:  "Documenting last/next/redo even further"
  From:  "M.J.T. Guy" <mjtg@cus.cam.ac.uk>, Tom Phoenix
         <rootbeer@teleport.com>
Msg-ID:  <E0yec2h-0000B9-00@taurus.cus.cam.ac.uk>,
         <Pine.GSO.3.96.980526111426.27437K-100000@user2.teleport.com>
 Files:  pod/perlfunc.pod

 Title:  "Documenting last/next/redo within continue block"
  From:  Tom Phoenix <rootbeer@teleport.com>
Msg-ID:  <Pine.GSO.3.96.980525214558.7133H-100000@user2.teleport.com>
 Files:  pod/perlfunc.pod

 Title:  "Document stat return in scalar context"
  From:  Mark-Jason Dominus <mjd@plover.com>
 Files:  pod/perlfunc.pod

 ------  EXTENSIONS  ------

 Title:  "Better LD_RUN_PATH handling on IRIX"
  From:  "W. Phillip Moore" <wpm@ms.com>
Msg-ID:  <199805212206.SAA07504@zappa.morgan.com>
 Files:  lib/ExtUtils/MM_Unix.pm

 Title:  "Dealing with <unistd.h> in POSIX and SunOS"
  From:  Andy Dougherty <doughera@lafcol.lafayette.edu>
Msg-ID:  <Pine.SUN.3.96.980527115256.26608F-100000@newton.phys>
 Files:  ext/POSIX/hints/sunos_4.pl hints/sunos_4_1.sh ext/POSIX/POSIX.xs

 ------  LIBRARY  ------

 Title:  "Fix FileHandle.pm example bug"
  From:  Daniel Grisinger <dgris@perrin.dimensional.com>
Msg-ID:  <Pine.LNX.3.96.980520194825.10845C-100000@perrin.dimensional.com>
 Files:  lib/FileHandle.pm

 Title:  "Add zero/negative $count docs for Benchmark.pm"
  From:  "M.J.T. Guy" <mjtg@cus.cam.ac.uk>
Msg-ID:  <E0ydEAr-0006NV-00@taurus.cus.cam.ac.uk>
 Files:  lib/Benchmark.pm

 ------  PORTABILITY - GENERAL  ------

 Title:  "Add test suite recommendations to Porting/patching.pod"
  From:  Daniel Grisinger <dgris@perrin.dimensional.com>
Msg-ID:  <Pine.LNX.3.96.980521162925.3568B-100000@perrin.dimensional.com>
 Files:  Porting/patching.pod

 ------  TESTS  ------

 Title:  "Fix looping bug in t/io/pipe.t"
  From:  "M.J.T. Guy" <mjtg@cus.cam.ac.uk>
Msg-ID:  <E0yc737-0006fB-00@taurus.cus.cam.ac.uk>
 Files:  t/io/pipe.t

p4raw-id: //depot/maint-5.004/perl@1038

+++ t/lib/h2ph.t
@@ -1,16 +1,28 @@
 #!./perl
 
+# quickie tests to see if h2ph actually runs and does more or less what is
+# expected
+
 BEGIN {
     chdir 't' if -d 't';
     @INC = '../lib';
 }
 
-use File::Compare; 
-
+use File::Compare;
 print "1..2\n";
-$ok = system("./perl -I../lib ../utils/h2ph -d. lib/h2ph.h");
-print(($ok == 0 ? "" : "not "), "ok 1\n");
-$ok = compare("lib/h2ph.ph", "lib/h2ph.pht");
-print(($ok == 0 ? "" : "not "), "ok 2\n");
-#system("diff -c lib/h2ph.pht lib/h2ph.ph 1>&2");
-unlink("lib/h2ph.ph");
+
+unless(-e '../utils/h2ph') {
+    print("ok 1\nok 2\n");
+    # i'll probably get in trouble for this :)
+} else {
+    # does it run?
+    $ok = system("./perl -I../lib ../utils/h2ph -d. lib/h2ph.h");
+    print(($ok == 0 ? "" : "not "), "ok 1\n");
+    
+    # does it work? well, does it do what we expect? :-)
+    $ok = compare("lib/h2ph.ph", "lib/h2ph.pht");
+    print(($ok == 0 ? "" : "not "), "ok 2\n");
+    
+    # cleanup - should this be in an END block?
+    unlink("lib/h2ph.ph");
+}
